feat(memory + sha3): fix wrong stack update (#150)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?
Stack isn't properly updated when storing word in memory and when
hashing a word
<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: N/A

## What is the new behavior?
Stack is properly updated with the latest version of it


## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->