Execute returns the whole stack instead of peek, tests uses parametrize (#165)
## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [x] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Currently, `kakarot.execute` only returns the peek of the stack. This is
not enough to insure that the execution is correct as the whole stack
needs to be tested.

## What is the new behavior?

Return the whole stack instead.
Tests have been updated accordingly and rewritten using
`pytest.mark.parametrize`.
The failing tests have been marked `skipped` and should be fixed in
another PR.

## Other information

We also need to return `storage` and `return_value` to battle test this
EVM, see returned value from [the
playground](https://www.evm.codes/playground?fork=grayGlacier)

The `coverage` module has been removed as this final coverage op is
carried out in the `starknet` fixture.