Add wrapper for web3 contract to work with kakarot (#267)
## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

When deploying a compiled solidity contract to kakarot, there is no easy
way to use the known abi to test target functions.

## What is the new behavior?

A wrapper of `web3.Contract` to be used after contract deployment that
allow for using the contract like `await contract.method_name(*args,
**kwargs)` handling both `call` and `execute` automatically.

This wrapper is itself used in a factory fixture called
`deploy_solidity_contract` to be used "as is" in test, e.g.:

```
async def test_deploy(
    self,
    deploy_solidity_contract: Callable,
):
    erc_20 = await deploy_solidity_contract("ERC20", "name", "symbol", 18, caller_address=1)
    name = await erc_20.name()
    assert name == "name"
    symbol = await erc_20.symbol()
    assert symbol == "symbol"
    decimals = await erc_20.decimals()
    assert decimals == 18
```

## Other information

There is probably a better way to integrate with `web3` but I didn't
want to spend to much time on a specific integration of one given web3
lib. This is more a first step to ease our testings with the probable
increase of solidity input files.

I could spend a bit more time to update the `constructor` attribute, as
well as parsing the output with the abi.