feat: Add 0xFD REVERT opcode (#277)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [x] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: #249 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Reoder SystemOperations opcodes
- Create 0xfd REVERT opcode
- Use in this initial implementation StarkNet's native revert
- Supports revert reason limited to the short string size
- Add unit test file for SystemOperation opcodes, create revert test

## Other information
In order to be fully compatible with EVM's revert we need to update the
behavior in the future to load the whole revert reason data based on the
`size` parameter passed in the stack. The current implementation is
limited by only supporting a short string as revert reason.

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->