bugfix: handle zero padded cases and handle dynamic gas in extcodecopy (#345) (#369)
<!--- Please provide a general summary of your changes in the title
above -->

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Cases where there was no match for a given account in the bytecode
registry would return context without updating the stack. It also wasn't
correctly returning zeros in memory for the cases where there was no
match

Resolves #345 

## What is the new behavior?

Now when there is no bytecode, zeros still written in memory for the
opcode `size` param at the desired `dest_offset`.

<!-- Please describe the behavior or changes that are being added by
this PR. -->

-
-
-

## Other information

I also added two of the necessary dimensions to compute the dynamic gas,
while still noting that the `address_access_cost` depends on determining
warm or cold access, which will be addressed in the future.

I created a util function `compute_minimum_word_size` that does the math
in uint256. This mitigates the case of the (length + 31) / 32 not
dividing evenly in felt arithmetic when the numerator is not a multiple
of the denominator, which would cause overflow. If that helper looks
could I can update the codebase to use it.