Kakarot Ethereum Account Abstraction (#422)
The Kakarot Ethereum Account Abstraction (KETHAA) is a set of two
contracts, a deployer and the account. The deployer is in charge of
deploying the accounts and the account manages the EVM transaction
verification and execution. For the moment only EIP-1559 transactions
are considered valid and no tx is executed. Few more things will be
added (see related issues). Not final state.

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Resolves #388 

## What is the new behavior?

Kakarot itself is not altered.

## Other information

Time spent on this PR:  10 days

Co-authored-by: kakarot CI <sayajin-labs@users.noreply.github.com>