Load revert reason short string (#431)
## Pull request type

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The revert opcode does not load the revert reason but only returns a
random number.

Resolves #416 

## What is the new behavior?

The reason is loaded as a bytes array. Only the short string is
propagated with the `error_attr`.

## Other information

Time spent on this PR:  1 day

We definitely need to fix REVERT (see #432) because this is close to a
nightmare in terms of hard fixing to have something ok for testing.