Fix usage of safe math (#450)
Replaced SafeMath arithmetic operations with uint256 operations. 

I spent two days, it was my first Cairo and EVM pr and used to
understand some details.

## Pull request type

Please check the type of change your PR introduces:

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The PR [#413](https://github.com/sayajin-labs/kakarot/pull/415) fixed
the usage of SafeMath on Multiplication and Addition, the rest of the
arithmetic operations are still using SafeMath.

Resolves #419

## What is the new behavior?

All arithmetic operations now use
[uint256](https://github.com/starkware-libs/cairo-lang/blob/master/src/starkware/cairo/common/uint256.cairo).

## Other information

According with the specs on the EVM we should'n use safe mode. The EVM
use "wrap-around" arithmetic, which means that overflow and underflow
conditions are not detected or handled in the same way as in traditional
computer architectures. Instead, the EVM wraps around to the opposite
end of the range of possible values.

In summary if we use SafeMath we would have incompatibility issues.

Co-authored-by: Francisco Strambini <franciscostrambini@Franciscos-MacBook-Air.local>
Co-authored-by: kakarot CI <sayajin-labs@users.noreply.github.com>