Feat/eoa_deploys_contract (#502)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 0.5d

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
Account contracts have to be deployed by interaction with kakarot and
can't be deployed via the EOA.

Resolves #491 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Sending a transaction (legacy or default) with the "to" address set to
"" will result in a contract deployment instead of a standard
transaction.
- An event containing the new contracts EVM address is emitted on
contract creation

## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

---------

Co-authored-by: Elias Tazartes <66871571+Eikix@users.noreply.github.com>