feat: EOA transfer token value  (#526)
<!--- Please provide a general summary of your changes in the title
above -->
Add the transfer of tokens in the `execute` method of the EOA.

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 1 days

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
Currently any transfer of value from an EOA isn't executed.

Resolves #466 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Transfer Kakarot native tokens during call to `execute` from EOA.
- Unit test