Fix/revert should not stop execution in sub context (#489)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 4.6 days

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Resolves #432 

## What is the new behavior?

- event emission opcodes push events to the execution context, which in
turn get emitted with a sys call only when an execution context is
stopped and not reverted
- create addresses are pushed to the execution context, which in turn
get erased when a context is stopped
- contract state prior to an operating context is monitored so we can
revert to it in the case that the operating context is reverted

## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->