feature: add `query_logs` method to kkrt wrapper (#540)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 0.5

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [X] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Revert tests were unable to assert that a given event emitted from a
subcontext was not emitted.

Resolves #<Issue number>

## What is the new behavior?

The kakarot wrapper now has a `query_logs` method that, when called with
- no args -> returns the raw logs of transaction.
- contract key -> returns all the events in logs that match the event
abi of the contract
- contract key x event name -> returns the events that match the
particular event name

## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->