gasprice should return txn constant (#548)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: .1

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Gasprice was returning a nonconstant value that was dependent on
execution.

Resolves #<Issue number>

## What is the new behavior?

Gasprice returns the amount of ether that the sender is willing to pay
per unit of gas to execute the transaction.


## Other information

I believe gasPrice is relayed from the rpc to the execution system, but
currently only `gas_limit` is parameterized at the entry point of
`execute_at_address`, so pragmaticly this opcode is always returning
zero.