fix: dependencies update (#559)
<!--- Please provide a general summary of your changes in the title
above -->
Updates necessary dependencies to be able to interact with the current
version of Starknet deployed on the testnet (v0.11).

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 0.75 day

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?
The current dependencies are outdated and prevent the interaction with
Starknet 0.11 deployed on the testnet.

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Resolves #558 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Starknet-py updated to version 0.15 which fixes failing interactions
with the testnet.
- Cairo-lang updated to the latest version 0.11
- Replace py-evm by blake2b-py 

## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->