feat: deploy account contract + execute at address utils (#550)
<!--- Please provide a general summary of your changes in the title
above -->
This PR provides python utilities in order to easily deploy a contract
on Kakarot and execute calldata on deployed contracts.

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 1 day

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
No script available to deploy a contract account on Kakarot or execute
calldata on a deployed contract.

Resolves N/A

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Provide a script to deploy a contract account or execute calldata on a
deployed contract.