fix: update scripts (#565)
<!--- Please provide a general summary of your changes in the title
above -->
An error caused the deployment script to crash when missing an evm
private key in the .env file.

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: N/A

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->
Error in script prevents CI deployment of contracts.

Resolves N/A

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Scripts are updated in order to allow deployment with missing evm
private key.