Feat/add katana docker container (#594)
<!--- Please provide a general summary of your changes in the title
above -->

The PR adds a Dockerfile to build a katana image with Kakarot already
deployed on it. The image is multi-arch and the pr also provides
integration of its build and publishing in the current github workflow.

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

We don't have a docker image of Katana with Kakarot already deployed on
it.

Resolves #589 #575 

## What is the new behavior?

provides a multi-arch docker image of katana with kakarot already
deployed on it.

## Other information

The containers of these images will also have a healthcheck, the
containers will become healthy, hence for testing, or building on top of
these containers, the health can be inspected.