Minor updates to deploy scripts after latest sharingan release (#602)
Time spent on this PR: 0.25

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

Two unrelated topics:

1. the chain_id enum has some hard-coded values that needs to be
maintained
1. the deploy scripts first deploy an OZ account in madara and
sharingan. This was done because no standard account was available on
these networks

## What is the new behavior?

Two unrelated updates:

1. the chain_id is now taken from the network itself so it's always the
right one
1. no more initial deployment of an account in madara and sharingan
because an OZ account has been added in the genesis state. The util is
not removed though because it is useful to have it ready in case we need
to create on the fly several accounts.

## Other information