Update .env.example for katana (#614)
Time spent on this PR: 0.01

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [x] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The default account for katana in the `.env.example` doesn't exist
anymore

## What is the new behavior?

Updated values

## Other information

While trying to deploy kakarot on katana, I was able to deploy a
`Counter` but not an ERC20, still facing the events issue. Not sure why
it would work for Counter and not ERC20 though. Just added in the
printed error the receipt's events.