Improve logging of starknet utils (#623)
Time spent on this PR: 0.1

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [x] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

The invoke/deploy/declare functions always log with a ✅ even though the
tx didn't succeed.
This "regression" has been introduced when the `wait_for_transaction`
has been updated to handle
"Transaction not found" by introducing a timeout.

Currently, at the end of the timeout, even if the tx is not
`ACCEPTED_ON_L2`, the logs show a misleading ✅.

## What is the new behavior?

Show ❌ when status is not ACCEPTED_ON_L2

## Other information

Usually, clients raise when a tx is not eventually `ACCEPTED_ON_L2`,
while we currently don't.
I added the emoji to make the logging more accurate without changing the
behavior. Actually I don't
know if we should raise, the problem coming from the fact that REJECTED
tx are not found neither in
RPCs. Say it differently, NOT_RECEIVED/PENDING/REJECTED return the same
"Transaction not found".