Use Gateway provider in testnet (#626)
Time spent on this PR: 0.3

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

We migrated to use only a RPC starknet provider. However, in the latest
Starknet upgrade, the RPC doesn't return
the same data as the Gateway (espc. transaction status, nonce).
After spending a bit of time to hack the RPC data, I decided to just put
the Gateway client for these networks where it is available.

## What is the new behavior?

Use Gateway for mainnet/testnet/testnet2

## Other information