Minor devX improvements (#641)
Time spent on this PR: 0.1

## Pull request type

Please check the type of change your PR introduces:

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

`chain_id` is fetch from the given RPC endpoint. Even though we now use
the Gateway for mainnet/testnet/testnet2,
we then still need an RPC endpoint just for this.

`fetch_deployments` is called when network is not `devnet` and raises if
no deployment is found. This is not
really necessary and was done so when we used only
devnet/testnet/mainnet.

## What is the new behavior?

`chain_id` is prefilled for testnet/testnet2/mainnet
`try/catch` around `fetch_deployments`