test_create_has_determistic_address tests address w/o state (#655)
<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 0.01

## Pull request type



- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

#649 changes the ability for testing to mock a nonce as a precondition
for an assert.

Resolves #654 

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- We separate the further testing of deterministic addresses by a range
of nonces to be done directly on `CreateHelper.get_create_address`,
instead of the previous way which depended on the ability to establish a
mock nonce value before running `SystemOperations.exec_create`

-
-

## Other information

<!-- Any other information that is important to this PR such as
screenshots of how the component looks before and after the change. -->

Currently there is a redundancy of only testing the zero case of nonce
for the `test_create`, we can factor out all nonce/address assertions in
`test_create` and have them done in a separate test,
`test_create_has_deterministic_address`

---------

Co-authored-by: kakarot CI <kkrt-labs@users.noreply.github.com>