fix: update events emissions logic (#672)
<!--- Please provide a general summary of your changes in the title
above -->
Currently we emit the evm address linked to the log as the last key of
the Starknet event. This causes issues when indexing events on Starknet,
as the evm address can be located as different places the in key array
depending on the amount of topics included in the log.

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR: 0.1 day

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?
Evm address is emitted as last key of the Starknet event, after all
topics.
<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Resolves #NA

## What is the new behavior?

<!-- Please describe the behavior or changes that are being added by
this PR. -->

- Evm address emitted as first key of the Starknet event, before any
topics