Add condition for zero return_data_len in finalization if-statement (#712)
Ensure that the finalization logic accounts for cases where
return_data_len is zero, to correctly branch into 'create' or 'call'
operations.

<!--- Please provide a general summary of your changes in the title
above -->

<!-- Give an estimate of the time you spent on this PR in terms of work
days. Did you spend 0.5 days on this PR or rather 2 days? -->

Time spent on this PR:

## Pull request type

<!-- Please try to limit your pull request to one type, submit multiple
pull requests if needed. -->

Please check the type of change your PR introduces:

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, renaming)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] Documentation content changes
- [ ] Other (please describe):

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Resolves #696 

## What is the new behavior?

per this suggestion:
https://github.com/kkrt-labs/kakarot/issues/696#issuecomment-1693405166

we add a check if the return data length is zero to decide whether the
execution context should be handed as create/call.

-
-
-